+++ tests/test_orca_parsers.py
@@ -1,11 +1,12 @@
 from pathlib import Path
 
 import pytest
-from qcdata import CalcType
+from qcdata import CalcType, Model, ProgramInput, Structure
 
 from qccodec.parsers.orca import (
     parse_energy,
     parse_gradient,
+    parse_gradient_hessian,
     parse_hessian,
     parse_natoms,
     parse_trajectory,
@@ -70,6 +71,15 @@ test_cases = [
         answer=gradients.water_revdsd,
     ),
     ParserTestCase(
+        name="Parse (absent) analytical gradient from hessian log",
+        parser=parse_gradient_hessian,
+        stdout=Path("water.hess.out"),
+        calctype=CalcType.hessian,
+        success=False,
+        decode_exc=False,
+        answer=None,
+    ),
+    ParserTestCase(
         name="Parse analytic hessian",
         parser=parse_hessian,
         stdout=Path("water.hess.out"),
@@ -88,6 +98,15 @@ test_cases = [
         extra_files=["water.numhess.hess"],
     ),
     ParserTestCase(
+        name="Parse hessian with single-column final block",
+        parser=parse_hessian,
+        stdout=Path("single_column.out"),
+        calctype=CalcType.hessian,
+        success=True,
+        answer=hessians.single_column,
+        extra_files=["single_column.hess"],
+    ),
+    ParserTestCase(
         name="Parse number of atoms water",
         parser=parse_natoms,
         stdout=Path("water.energy.out"),
@@ -104,6 +123,36 @@ test_cases = [
         answer=trajectories.trajectory,
         clear_registry=False,
         extra_files=["water.opt_trj.xyz"],
+    ),
+    ParserTestCase(
+        name="Parse trajectory (m=2)",
+        parser=parse_trajectory,
+        stdout=Path("ch3.opt.out"),
+        calctype=CalcType.optimization,
+        success=True,
+        answer=trajectories.trajectory_ch3,
+        clear_registry=False,
+        program_input=ProgramInput(
+            structure=Structure(
+                symbols=["C", "H", "H", "H"],
+                geometry=[
+                    [
+                        2.2960172429784643e-07,
+                        -5.47453658675606e-07,
+                        -0.13371279750931814,
+                    ],
+                    [1.20598062158438, 1.4745349158139784, 0.5311114980010905],
+                    [-1.879974623959173, 0.3071463406992686, 0.5311113957669071],
+                    [0.6739912652954723, -1.7816747322337971, 0.5311139136379973],
+                ],
+                charge=0,
+                multiplicity=2,
+            ),
+            model=Model(method="xtb"),
+            calctype=CalcType.optimization,
+        ),
+        extra_files=["ch3.opt_trj.xyz"],
+        extra_files_names=["orca_trj.xyz"],
     ),
 ]